+++ pmd-core/src/main/java/net/sourceforge/pmd/util/StringUtil.java
@@ -5,6 +5,7 @@
 package net.sourceforge.pmd.util;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 
@@ -19,13 +20,9 @@ import org.apache.commons.lang3.StringUtils;
 public final class StringUtil {
 
     private static final String[] EMPTY_STRINGS = new String[0];
-    private static final boolean SUPPORTS_UTF8 = "yes"
-        .equals(System.getProperty("net.sourceforge.pmd.supportUTF8", "no"));
-
 
     private StringUtil() {
     }
-
 
     /**
      * Formats a double to a percentage, keeping {@code numDecimal} decimal places.
@@ -77,7 +74,9 @@ public final class StringUtil {
      * @param prefixes
      *
      * @return boolean
+     * @deprecated {@see StringUtils#startsWithAny(CharSequence, CharSequence...)}
      */
+    @Deprecated
     public static boolean startsWithAny(String text, String... prefixes) {
 
         for (String prefix : prefixes) {
@@ -135,7 +134,9 @@ public final class StringUtil {
      * @param value String
      *
      * @return boolean
+     * @deprecated {@see StringUtils#isNotBlank(CharSequence)}
      */
+    @Deprecated
     public static boolean isNotEmpty(String value) {
         return !isEmpty(value);
     }
@@ -149,7 +150,9 @@ public final class StringUtil {
      * @param value String to test
      *
      * @return <code>true</code> if the value is empty, <code>false</code> otherwise.
+     * @deprecated {@see StringUtils#isBlank(CharSequence)}
      */
+    @Deprecated
     public static boolean isEmpty(String value) {
         return StringUtils.isBlank(value);
     }
@@ -161,7 +164,9 @@ public final class StringUtil {
      * @param value String to test
      *
      * @return True if the argument is null or the empty string
+     * @deprecated {@see StringUtils#isEmpty(CharSequence)}
      */
+    @Deprecated
     public static boolean isMissing(String value) {
         return StringUtils.isEmpty(value);
     }
@@ -176,6 +181,7 @@ public final class StringUtil {
      *
      * @return boolean
      */
+    @Deprecated
     public static boolean areSemanticEquals(String a, String b) {
 
         if (a == null) {
@@ -195,7 +201,9 @@ public final class StringUtil {
      * @param newString String
      *
      * @return String
+     * @deprecated {@see StringUtils#replace(String, String, String)}
      */
+    @Deprecated
     public static String replaceString(final String original, final String oldString, final String newString) {
         int index = original.indexOf(oldString);
         if (index < 0) {
@@ -214,22 +222,6 @@ public final class StringUtil {
             return buf.toString();
         }
     }
-
-
-    /**
-     * Appends to a StringBuilder the String src where non-ASCII and XML special
-     * chars are escaped.
-     *
-     * @param buf The destination XML stream
-     * @param src The String to append to the stream
-     *
-     * @deprecated use {@link #appendXmlEscaped(StringBuilder, String, boolean)} instead
-     */
-    @Deprecated
-    public static void appendXmlEscaped(StringBuilder buf, String src) {
-        appendXmlEscaped(buf, src, SUPPORTS_UTF8);
-    }
-
 
     /**
      * @param buf
@@ -291,29 +283,15 @@ public final class StringUtil {
         return s;
     }
 
-
-    /**
-     * @param string String
-     *
-     * @return String
-     *
-     * @deprecated Use StringEscapeUtils#escapeHtml4 instead
-     */
-    @Deprecated
-    public static String htmlEncode(String string) {
-        String encoded = replaceString(string, '&', "&amp;");
-        encoded = replaceString(encoded, '<', "&lt;");
-        return replaceString(encoded, '>', "&gt;");
-    }
-
-
     /**
      * @param original  String
      * @param oldChar   char
      * @param newString String
      *
      * @return String
+     * @deprecated {@see StringUtils#replace(String, String, String)} or {@see StringUtils#replaceChars(String, char, char)}
      */
+    @Deprecated
     public static String replaceString(final String original, char oldChar, final String newString) {
         int index = original.indexOf(oldChar);
         if (index < 0) {
@@ -340,13 +318,13 @@ public final class StringUtil {
      * and serves as a replacement for String.split() for JDK1.3 that doesn't
      * have it.
      *
-     * FIXME - we're on JDK 1.4 now, can we replace this with String.split?
-     *
      * @param source    String
      * @param delimiter char
      *
      * @return String[]
+     * @deprecated {@see StringUtils#split(String, char)}
      */
+    @Deprecated
     public static String[] substringsOf(String source, char delimiter) {
 
         if (source == null || source.length() == 0) {
@@ -392,7 +370,9 @@ public final class StringUtil {
      * @param separator char
      *
      * @return String[]
+     * @deprecated {@see StringUtils#split(String, String)}
      */
+    @Deprecated
     public static String[] substringsOf(String str, String separator) {
 
         if (str == null || str.length() == 0) {
@@ -424,7 +404,9 @@ public final class StringUtil {
      * @param sb        StringBuffer
      * @param iter      Iterator
      * @param separator String
+     * @deprecated {@see StringUtils#join(Iterator, String)}
      */
+    @Deprecated
     public static void asStringOn(StringBuffer sb, Iterator<?> iter, String separator) {
 
         if (!iter.hasNext()) {
@@ -447,7 +429,9 @@ public final class StringUtil {
      * @param sb        StringBuilder
      * @param items     Object[]
      * @param separator String
+     * @deprecated {@see StringUtils#join(Iterable, String)}
      */
+    @Deprecated
     public static void asStringOn(StringBuilder sb, Object[] items, String separator) {
 
         if (items == null || items.length == 0) {
@@ -557,12 +541,14 @@ public final class StringUtil {
      * @param length The desired minimum length of the resulting padded String
      *
      * @return The resulting left padded String
+     * @deprecated {@see StringUtils#leftPad(String, int)}
      */
+    @Deprecated
     public static String lpad(String s, int length) {
         String res = s;
         if (length - s.length() > 0) {
             char[] arr = new char[length - s.length()];
-            java.util.Arrays.fill(arr, ' ');
+            Arrays.fill(arr, ' ');
             res = new StringBuilder(length).append(arr).append(s).toString();
         }
         return res;